refactor: change the file name from Auth to Login controller
